requirements.txt
Django==4.1.7
djangorestframework==3.13.1
djangorestframework-simplejwt~=5.2.2

psycopg2~=2.9.6
psycopg2-binary==2.9.6

celery==5.2.7
redis==4.3.1

requests~=2.28.2
python-dotenv~=1.0.0
